Moved UI cancelling code to PatchingUtilities
